SELinux: do all flags twiddling in one place
Currently we set the initialize and seclabel flag in one place.  Do some
unrelated printk then we unset the seclabel flag.  Eww.  Instead do the flag
twiddling in one place in the code not seperated by unrelated printk.  Also
don't set and unset the seclabel flag.  Only set it if we need to.

Change-Id: I6c5bf9ddd32084a9d52ee640bd2581499de62ac0
Signed-off-by: Eric Paris <eparis@redhat.com>